Few but important changes
Advance MPU6050 configuration with an integral Oppy driver.

+++ PeripheralDrivers/Src/I2CDriver.c
@@ -11,7 +11,7 @@
 #include "PLLDriver.h"
 
 
-uint64_t freqClock1 = 0;
+uint32_t freqClock1 = 0;
 
 /*
  * Recordar que se debe configurar los pines para el I2C (SDA y SCL),
